requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
Brotli==1.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dataclasses-json==0.6.7
Deprecated==1.2.18
dj-database-url==2.3.0
Django==5.2.1
djangorestframework==3.16.0
dotenv==0.9.9
drf-spectacular==0.28.0
dune_client==1.7.9
frozenlist==1.6.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
ndjson==0.3.1
packaging==25.0
passlib==1.7.4
propcache==0.3.1
protobuf==6.31.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20250516
types-PyYAML==6.0.12.20250516
types-requests==2.32.0.20250515
types-setuptools==80.7.0.20250516
typing-inspect==0.9.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
whitenoise==6.9.0
wrapt==1.17.2
yarl==1.20.0
